PSP2: Add sharp-bilinear-simple shader back in.
This is a good shader for  users like me who hate scanline, but
want a sharp picture without pixelwobble. The shader looks great
with some games like Air Gallet with zero slow-down. It gives a
sharper picture than just 'bilinear+none', and the pixels remain
square.

I suppose it shouldn't hurt to have it as an option, so I am
adding it back in.